modularize dataset creation args processing
