+++ hsds/util/chunkUtil.py
@@ -11,230 +11,6 @@ DEFAULT_TYPE_SIZE = 128  # Type size case when it is variable
 PRIMES = [29, 31, 37, 41, 43, 47, 53, 59, 61, 67]  # for chunk partitioning
 
 
-def getChunkSize(layout, type_size):
-    """Return chunk size given layout.
-    i.e. just the product of the values in the list.
-    """
-    if type_size == "H5T_VARIABLE":
-        type_size = DEFAULT_TYPE_SIZE
-
-    chunk_size = type_size
-    for n in layout:
-        if n <= 0:
-            raise ValueError("Invalid chunk layout")
-        chunk_size *= n
-    return chunk_size
-
-
-def get_dset_size(shape_json, typesize):
-    """Return the size of the dataspace.  For
-    any unlimited dimensions, assume a value of 1.
-    (so the return size will be the absolute minimum)
-    """
-    if shape_json is None or shape_json["class"] == "H5S_NULL":
-        return None
-    if shape_json["class"] == "H5S_SCALAR":
-        return typesize  # just return size for one item
-    if typesize == "H5T_VARIABLE":
-        typesize = DEFAULT_TYPE_SIZE  # just take a guess at the item size
-    dset_size = typesize
-    shape = shape_json["dims"]
-    rank = len(shape)
-
-    for n in range(rank):
-        if shape[n] == 0:
-            # extendable extent with value of 0
-            continue  # assume this is one
-        dset_size *= shape[n]
-    return dset_size
-
-
-def expandChunk(
-    layout, typesize, shape_json, chunk_min=CHUNK_MIN, layout_class="H5D_CHUNKED"
-):
-    """Compute an increased chunk shape with a size in bytes greater than chunk_min."""
-    if shape_json is None or shape_json["class"] == "H5S_NULL":
-        return None
-    if shape_json["class"] == "H5S_SCALAR":
-        return (1,)  # just enough to store one item
-
-    layout = list(layout)
-    log.debug(f"expandChunk layout: {layout} typesize: {typesize}")
-    dims = shape_json["dims"]
-    rank = len(dims)
-    extendable_dims = 0  # number of dimensions that are extenable
-    maxdims = None
-    if "maxdims" in shape_json:
-        maxdims = shape_json["maxdims"]
-        for n in range(rank):
-            if maxdims[n] == 0 or maxdims[n] > dims[n]:
-                extendable_dims += 1
-
-    dset_size = get_dset_size(shape_json, typesize)
-    if dset_size <= chunk_min and extendable_dims == 0:
-        # just use the entire dataspace shape as one big chunk
-        return tuple(dims)
-
-    chunk_size = getChunkSize(layout, typesize)
-    if chunk_size >= chunk_min:
-        return tuple(layout)  # good already
-    while chunk_size < chunk_min:
-        # just adjust along extendable dimensions first
-        old_chunk_size = chunk_size
-        for n in range(rank):
-            dim = rank - n - 1  # start from last dim
-
-            if extendable_dims > 0:
-                if maxdims[dim] == 0:
-                    # infinitely extendable dimensions
-                    layout[dim] *= 2
-                    chunk_size = getChunkSize(layout, typesize)
-                    if chunk_size > chunk_min:
-                        break
-                elif maxdims[dim] > layout[dim]:
-                    # can only be extended so much
-                    layout[dim] *= 2
-                    if layout[dim] >= dims[dim]:
-                        layout[dim] = maxdims[dim]  # trim back
-                        extendable_dims -= 1  # one less extenable dimension
-
-                    chunk_size = getChunkSize(layout, typesize)
-                    if chunk_size > chunk_min:
-                        break
-                    else:
-                        pass  # ignore non-extensible for now
-            else:
-                # no extendable dimensions
-                if dims[dim] > layout[dim]:
-                    # can expand chunk along this dimension
-                    layout[dim] *= 2
-                    if layout[dim] > dims[dim]:
-                        layout[dim] = dims[dim]  # trim back
-                    chunk_size = getChunkSize(layout, typesize)
-                    if chunk_size > chunk_min:
-                        break
-                else:
-                    pass  # can't extend chunk along this dimension
-        if chunk_size <= old_chunk_size:
-            # stop iteration if we haven't increased the chunk size
-            log.debug("stopping expandChunk iteration")
-            break
-        elif chunk_size > chunk_min:
-            break  # we're good
-        else:
-            pass  # do another round
-    return tuple(layout)
-
-
-def shrinkChunk(layout, typesize, chunk_max=CHUNK_MAX, layout_class="H5D_CHUNKED"):
-    """Compute a reduced chunk shape with a size in bytes less than chunk_max."""
-    layout = list(layout)
-    chunk_size = getChunkSize(layout, typesize)
-    if chunk_size <= chunk_max:
-        return tuple(layout)  # good already
-    log.debug(f"shrinkChunk layout: {layout} typesize: {typesize}")
-    rank = len(layout)
-
-    while chunk_size > chunk_max:
-        # just adjust along extendable dimensions first
-        old_chunk_size = chunk_size
-        for dim in range(rank):
-            if layout[dim] > 1:
-                # tricky way to do  x // 2 with ceil
-                layout[dim] = -(-layout[dim] // 2)
-                chunk_size = getChunkSize(layout, typesize)
-                if chunk_size <= chunk_max:
-                    break
-            else:
-                pass  # can't shrink chunk along this dimension
-        if chunk_size >= old_chunk_size:
-            # reality check to see if we'll ever break out of the while loop
-            log.warning("Unexpected error in shrink_chunk")
-            break
-        elif chunk_size <= chunk_max:
-            break  # we're good
-        else:
-            pass  # do another round
-    return tuple(layout)
-
-
-def guessChunk(shape_json, typesize):
-    """Guess an appropriate chunk layout for a dataset, given its shape and
-    the size of each element in bytes.  Will allocate chunks only as large
-    as MAX_SIZE.  Chunks are generally close to some power-of-2 fraction of
-    each axis, slightly favoring bigger values for the last index.
-
-    Undocumented and subject to change without warning.
-    """
-    if shape_json is None or shape_json["class"] == "H5S_NULL":
-        return None
-    if shape_json["class"] == "H5S_SCALAR":
-        return (1,)  # just enough to store one item
-
-    if "maxdims" in shape_json:
-        shape = shape_json["maxdims"]
-    else:
-        shape = shape_json["dims"]
-
-    if typesize == "H5T_VARIABLE":
-        typesize = 128  # just take a guess at the item size
-
-    # For unlimited dimensions we have to guess. use 1024
-    shape = tuple((x if x != 0 else 1024) for i, x in enumerate(shape))
-
-    return shape
-
-
-def getContiguousLayout(
-    shape_json, item_size, chunk_min=1000 * 1000, chunk_max=4 * 1000 * 1000
-):
-    """
-    create a chunklayout for datasets use continguous storage.
-    """
-    if not isinstance(item_size, int):
-        msg = "ContiguousLayout can only be used with fixed-length types"
-        raise ValueError(msg)
-    if chunk_max < chunk_min:
-        raise ValueError("chunk_max cannot be less than chunk_min")
-    if shape_json is None or shape_json["class"] == "H5S_NULL":
-        return None
-    if shape_json["class"] == "H5S_SCALAR":
-        return (1,)  # just enough to store one item
-    dims = shape_json["dims"]
-    rank = len(dims)
-    if rank == 0:
-        raise ValueError("rank must be positive for Contiguous Layout")
-    for dim in dims:
-        if dim < 0:
-            raise ValueError("extents must be positive for Contiguous Layout")
-        if dim == 0:
-            # datashape with no elements, just return dims as layout
-            return dims
-
-    nsize = item_size
-    layout = [
-        1,
-    ] * rank
-
-    for i in range(rank):
-        dim = rank - i - 1
-        extent = dims[dim]
-        if extent * nsize < chunk_max:
-            # just use the full extent as layout
-            layout[dim] = extent
-            nsize *= extent
-        else:
-            n = extent
-            while n > 1:
-                n = -(-n // 2)  # use negatives so we round up on odds
-                if n * nsize < chunk_max:
-                    break
-            layout[dim] = n
-            break  # just use 1's for the rest of the layout
-
-    return layout
-
-
 def frac(x, d):
     """
     Utility func -- Works like fractional div, but returns ceiling
